fix: address cross-platform quality failures
