+++ docs/architecture/issue-1000-manual-release-qa-preflight.md
@@ -122,7 +122,7 @@ missing MCP.
 | --- | --- |
 | Scenario and host exposure | RAES parsing/planning, `raes_realization_values.published_ports()`, `DeploymentPublishedPort`, `_compose_port_realization`, generated Compose, and `raes_runtime_observation._observe_published_ports()` remain the declaration-to-readback chain. Do not infer ports from MCP JSON or the retired proxy. |
 | Lifecycle and final outcome | `orchestrate_lab_start()`, `_LabStartContext`, `_LAB_START_STEPS`, `LabResult`, `StartupOutcome`, `StartupDiagnostic`, `_emit_diagnostic()`, and the issue #905 lifecycle lock remain the sole public-start contract. |
-| Project inventory | `DeploymentBackend`, local/SSH Compose backends, `_compose_queries`, `LabStatus`, `ProjectRuntimePresence`, project labels, and the existing `container_inspect()` detail seam own enumeration and state. Consolidate their query/parsing logic instead of adding another inventory DTO or raw command in core/CLI. |
+| Project inventory | `DeploymentBackend`, local/SSH Compose backends, `_compose_project_inventory`, `LabStatus`, `ProjectRuntimePresence`, project labels, and the existing `container_inspect()` detail seam own enumeration and state. Consolidate their query/parsing logic instead of adding another inventory DTO or raw command in core/CLI. |
 | User projections | `cli.lab._emit_status_text()`, `api.schemas.ContainerInfo` / `LabStatusResponse`, the API lab router, and `web/src/lib/types.ts` project core status. They do not recalculate ownership or readiness. |
 | Snapshot and persistence | `capture_snapshot()`, `RangeSnapshot.to_dict()`, `LocalRunStore` redacting JSON/JSONL writes, run provenance, and `aptl runs` remain the machine evidence boundary. Manual evidence indexes may reference these artifacts; they do not define another experiment/run schema. |
 | Live validation | `validate_live_deployment()`, `_live_gate_checks`, `_live_gate_readiness`, `LiveGateCheck`, `LiveGateReport`, and the existing scenario-verifier plugin keep their current failure taxonomy. |
